Mais alteração

+++ src/menu.h
@@ -11,8 +11,8 @@ using std::string;
 
 void menu();
 void init_db(string);
-void adicionarDespesa(string nome, int valor);
+void adicionarDespesa(string nome, double valor);
 void visualizarResumo();
 void removerDespesa(int index);
-void atualizarDespesa(int index, string new_name, int new_valor);
+void atualizarDespesa(int index, string new_name, double new_valor);
 #endif // MENU_H
